+++ test/test_throws.jl
@@ -7,9 +7,6 @@
         @test_throws DimensionMismatch GaussianState(basis1, [1.0, 2.0, 3.0], [3.0 4.0; 5.0 6.0])
         @test_throws DimensionMismatch GaussianChannel(basis1, [1.0, 2.0, 3.0], [3.0 4.0; 5.0 6.0], [3.0 4.0; 5.0 6.0])
         @test_throws DimensionMismatch GaussianChannel(basis1, [1.0, 2.0], [3.0 4.0; 5.0 6.0], [3.0 4.0 4.0; 5.0 6.0 4.0])
-        vac = vacuumstate(basis1)
-        vacs = vac ⊗ vac ⊗ vac ⊗ vac
-        @test_throws DimensionMismatch Generaldyne(vacs, vac ⊗ vac, [2, 4, 5])
     end
 
     @testset "action throws" begin